+++ app/Http/Controllers/OrderSpecialStatusController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Auth;
 use Alert;
+use App\Models\FacebookOrder;
 use App\Models\OrderSpecialStatus;
 use Image;
 use File;
@@ -91,6 +92,11 @@ class OrderSpecialStatusController extends Controller {
     public function destroy($id) {
         if (auth()->user()->can('size.create')) {
             $status = OrderSpecialStatus::find($id);
+
+            if (FacebookOrder::where('special_status_id', $id)->exists()) {
+                Alert::toast("Status can't be deleted because there is orders under this status !", 'warning');
+                return redirect()->route('fos.special_status.index');
+            };
 
             if (!is_null($status)) {
                 $status->delete();
